Refactored completions classes for better readability.
