+++ src/main/java/com/tabnine/intellij/completions/InsertNothingLookupElement.java
@@ -1,4 +1,4 @@
-package com.tabnine;
+package com.tabnine.intellij.completions;
 
 import com.intellij.codeInsight.lookup.LookupElement;
 import com.intellij.codeInsight.lookup.LookupElementDecorator;
