+++ src/main/java/com/tabnine/intellij/completions/TabNineCompletionContributor.java
@@ -1,4 +1,4 @@
-package com.tabnine;
+package com.tabnine.intellij.completions;
 
 import com.intellij.codeInsight.completion.*;
 import com.intellij.codeInsight.lookup.*;
